test, scenarios: use LNChannel class to compare
